data displayed in table and sortable
