+++ src/components/DataTypeTable/index.js
@@ -1,6 +1,5 @@
-import Type from './Type';
 import TypesList from './TypesList';
 
 // notice we're building out a 'package' of reusables here and exporting them as an object of component properties.
 // to use this, simply `import {foo, bar, baz} from '<path-to-this-directory>/ReusableComponents';`
-export { Type, TypesList };
+export { TypesList };
